Move Color to the math folder (#2804)
