+++ src/deprecated.js
@@ -1,9 +1,9 @@
 import { revision, version } from './core/core.js';
 import { string } from './core/string.js';
-import { Color } from './core/color.js';
 import { Timer, now } from "../src/core/time";
 
 import { math } from './math/math.js';
+import { Color } from './math/color.js';
 import { Mat4 } from './math/mat4.js';
 import { Vec2 } from './math/vec2.js';
 import { Vec3 } from './math/vec3.js';
