+++ src/font/canvas-font.js
@@ -1,6 +1,7 @@
 import { string } from '../core/string.js';
-import { Color } from '../core/color.js';
 import { EventHandler } from '../core/event-handler.js';
+
+import { Color } from '../math/color.js';
 
 import {
     ADDRESS_CLAMP_TO_EDGE,
