+++ src/framework/components/button/data.js
@@ -1,5 +1,4 @@
-import { Color } from '../../../core/color.js';
-
+import { Color } from '../../../math/color.js';
 import { Vec4 } from '../../../math/vec4.js';
 
 import { BUTTON_TRANSITION_MODE_TINT } from './constants.js';
