+++ src/framework/components/camera/system.js
@@ -1,5 +1,4 @@
-import { Color } from '../../../core/color.js';
-
+import { Color } from '../../../math/color.js';
 import { Vec4 } from '../../../math/vec4.js';
 
 import { Component } from '../component.js';
