+++ src/framework/components/sprite/component.js
@@ -1,6 +1,5 @@
-import { Color } from '../../../core/color.js';
-
 import { math } from '../../../math/math.js';
+import { Color } from '../../../math/color.js';
 import { Vec2 } from '../../../math/vec2.js';
 import { Vec4 } from '../../../math/vec4.js';
