+++ src/framework/components/sprite/system.js
@@ -1,4 +1,4 @@
-import { Color } from '../../../core/color.js';
+import { Color } from '../../../math/color.js';
 
 import {
     CULLFACE_NONE,
