+++ src/scene/layer.js
@@ -1,5 +1,6 @@
 import { hashCode } from '../core/hash.js';
-import { Color } from '../core/color.js';
+
+import { Color } from '../math/color.js';
 
 import { CLEARFLAG_COLOR, CLEARFLAG_DEPTH, CLEARFLAG_STENCIL } from '../graphics/constants.js';
