+++ src/scene/light.js
@@ -1,6 +1,5 @@
-import { Color } from '../core/color.js';
-
 import { math } from '../math/math.js';
+import { Color } from '../math/color.js';
 import { Mat4 } from '../math/mat4.js';
 import { Vec2 } from '../math/vec2.js';
 import { Vec3 } from '../math/vec3.js';
