+++ src/scene/lightmapper.js
@@ -1,7 +1,7 @@
 import { now } from '../core/time.js';
-import { Color } from '../core/color.js';
 
 import { math } from '../math/math.js';
+import { Color } from '../math/color.js';
 import { Vec3 } from '../math/vec3.js';
 
 import { BoundingBox } from '../shape/bounding-box.js';
