+++ src/scene/materials/standard-material.js
@@ -1,5 +1,4 @@
-import { Color } from '../../core/color.js';
-
+import { Color } from '../../math/color.js';
 import { Vec2 } from '../../math/vec2.js';
 import { Vec4 } from '../../math/vec4.js';
