+++ src/script/script-attributes.js
@@ -1,5 +1,4 @@
-import { Color } from '../core/color.js';
-
+import { Color } from '../math/color.js';
 import { Curve } from '../math/curve.js';
 import { CurveSet } from '../math/curve-set.js';
 import { Vec2 } from '../math/vec2.js';
