+++ src/xr/xr-light-estimation.js
@@ -1,6 +1,6 @@
-import { Color } from '../core/color.js';
 import { EventHandler } from '../core/event-handler.js';
 
+import { Color } from '../math/color.js';
 import { Mat4 } from '../math/mat4.js';
 import { Quat } from '../math/quat.js';
 import { Vec3 } from '../math/vec3.js';
